% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Check Hoot Owl Hoot! game}
\usage{
check(game)
}
\arguments{
\item{game}{A game object}
}
\value{
A character string with values win, loss, or in progress.
}
\description{
Game can be a win, loss, or in progress.
}
\examples{
game <- setup()
check(game)

}
\seealso{
\link{\code{setup}}
}

