% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_card.R
\name{check_card}
\alias{check_card}
\title{Check the card the player tries to play}
\usage{
check_card(card, game)
}
\arguments{
\item{card}{A string indicating the card a player is trying to play}

\item{game}{A current game state}
}
\value{
Invisibly TRUE
}
\description{
Check the card the player tries to play
}

