% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_owl.R
\name{check_owl}
\alias{check_owl}
\title{Check to make sure owl is valid}
\usage{
check_owl(owl, game)
}
\arguments{
\item{owl}{An integer specifying an owl}

\item{game}{A current game state}
}
\value{
Invisibly TRUE
}
\description{
Check to make sure owl is valid
}

