% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_and_draw.R
\name{discard_and_draw}
\alias{discard_and_draw}
\title{Discard a card and draw a new one}
\usage{
discard_and_draw(card, game)
}
\arguments{
\item{card}{A string indicating the card}

\item{game}{A current game state}
}
\value{
An updated game state
}
\description{
Discard a card and draw a new one
}
\seealso{
\link{\code{play_card}}
}

