% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_owl.R
\name{move_owl}
\alias{move_owl}
\title{Move an owl to specified location}
\usage{
move_owl(card, owl, game)
}
\arguments{
\item{card}{A string specifying a card}

\item{owl}{An integer specifying an owl}

\item{game}{A current game state}
}
\value{
An updated game state
}
\description{
Move an owl to specified location
}

