% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_player.R
\name{next_player}
\alias{next_player}
\title{Updates the game state to the next player}
\usage{
next_player(game)
}
\arguments{
\item{game}{A current game state}
}
\value{
An updated game
}
\description{
Updates the game state to the next player
}
\seealso{
\link{\code{play}}
}

