% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_card.R
\name{play_card}
\alias{play_card}
\title{Play a card}
\usage{
play_card(card, game, owl = NULL)
}
\arguments{
\item{card}{A character string indicating a card}

\item{game}{A game state}

\item{owl}{An integer specifying which owl to move. Owl #1 is closest to the
end.}
}
\value{
An updated game state
}
\description{
Play a card
}
\examples{
game <- setup_game()
game <- play_card(game$player$player1[1])

}
\seealso{
\code{\link{setup_game}}
}

