% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_game.R
\name{plot_game}
\alias{plot_game}
\title{Plot a current game state}
\usage{
plot_game(game)
}
\arguments{
\item{game}{A current game state}
}
\description{
Plot a current game state
}
\examples{
game <- setup_game()
plot_game(game)

}

