% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_experiments.R
\name{run_experiment}
\alias{run_experiment}
\title{Run a Hoot Owl Hoot! experiment}
\usage{
run_experiment(n_players, n_owls, strategy, n_cards_per_player, verbosity)
}
\arguments{
\item{n_players}{An integer specifying the number of players}

\item{n_owls}{An integer specifying the number of owls}

\item{strategy}{A function specifying a strategy for all players to use or a
list of length \code{n_players} where element i specifies a strategy for
player i. See details for more information.}

\item{n_cards_per_player}{An integer specifying the number of cards for each
player.}

\item{verbosity}{An integer specifying how much detail to print out (0=none).}
}
\value{
A data.frame with end game statistics
}
\description{
Simulate a Hoot Owl Hoot! game and record whether the game was
a win or loss, how many cards were played, and how many suns were played.
}
\examples{
run_experiment(n_players = 2, n_owls = 2, strategy = strategy_random)

}
\seealso{
\code{\link{simulate_game}}
}

