% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_experiments.R
\name{run_experiments}
\alias{run_experiments}
\title{Run a series of Hoot Owl Hoot! experiments}
\usage{
run_experiments(n_reps, n_players, n_owls, strategy = strategy_random,
  n_cards_per_player = 3, verbosity = 0)
}
\arguments{
\item{n_reps}{An integer indicating how many reps of each combination to run.}

\item{n_players}{A vector of integers specifying the number of players}

\item{n_owls}{A vector of integers specifying the number of owls}

\item{strategy}{A function specifying a strategy for all players to use or a
list of length \code{n_players} where element i specifies a strategy for
player i. The same strategy will be used for all games.}

\item{n_cards_per_player}{A vector of integers specifying the number of cards
for each player.}

\item{verbosity}{An integer specifying how much detail to print out (0=none).}
}
\description{
This function will simulate a bunch of Hoot Owl Hoot! games with a set
strategy and record the end game status of those games.
}
\examples{
run_experiments(n_reps = 2, n_players = 2:3, n_owls = 1:2,
                strategy = strategy_random,
                n_cards_per_player = 2:3,
                verbosity = 0)

}
\seealso{
\code{\link{run_experiment}}
}

