% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup Hoot Owl Hoot! Game}
\usage{
setup(n_players = 2, n_owls = 3, n_cards_per_player = 3)
}
\arguments{
\item{n_players}{An integer (2-4) specifying the number of players}

\item{n_owls}{An integer (1-6) specifying the number of owls}

\item{n_cards_per_player}{An integer specifying the number of cards per
player.}
}
\value{
A list with elements deck, player, board, sun, and discard.
  \code{deck} contains the shuffled cards
  \code{player} is a n_cards_per_player x n_players matrix with the in hand
    cards
  \code{board} a \code{data.frame} containing the spaces on the board and
    whether those spaces are occupied
  \code{sun} an integer specifying where the sun is (0 is a loss)
  \code{discard} the cards in the discard pile (initially NULL)
}
\description{
Setup Hoot Owl Hoot! Game
}
\examples{
board <- setup(n_players = 3, n_owls = 4)

}

