% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle cards}
\usage{
shuffle(cards)
}
\arguments{
\item{cards}{A character vector of cards}
}
\value{
A character vector of cards
}
\description{
Shuffle cards
}
\examples{
discard = c("this","that")
shuffle(discard)

}

