% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_game.R
\name{simulate_game}
\alias{simulate_game}
\title{Simulate a Hoot Owl Hoot! game}
\usage{
simulate_game(n_players, n_owls, strategy, n_cards_per_player = 3,
  verbose = 0)
}
\arguments{
\item{n_players}{An integer specifying the number of players}

\item{n_owls}{An integer specifying the number of owls}

\item{strategy}{A function specifying a strategy for all players to use or a
list of length \code{n_players} where element i specifies a strategy for
player i. See details for more information.}

\item{n_cards_per_player}{An integer specifying the number of cards for each
player.}

\item{verbosity}{An integer specifying how much detail to print out (0=none).}
}
\value{
A game end state
}
\description{
Simulate a Hoot Owl Hoot! game
}
\details{
A strategy is a function taking a \code{game} and returning a list
  with elements \code{card} and \code{owl} for the next play. See
  \link{\code{strategy_random}} for an example.
}
\examples{
simulate_game(2,2, strategy_random)

}
\seealso{
\link{\code{strategy_random}}
}

