% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_last_owl_farthest.R
\name{strategy_last_owl_farthest}
\alias{strategy_last_owl_farthest}
\title{Play a last-owl-farthest strategy}
\usage{
strategy_last_owl_farthest(game)
}
\arguments{
\item{game}{A current game state}
}
\value{
list with these elements:
  \code{card} string indicating what card to play
  \code{owl} an integer indicating the owl to move
}
\description{
Play a sun if you have a sun, otherwise move the last owl to the location
that moves them the farthest.
}
\examples{
game <- setup_game()
strategy_last_owl_farthest(game)
}
\seealso{
\code{\link{strategy_random}}
}

