% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_last_owl_random.R
\name{strategy_last_owl_random}
\alias{strategy_last_owl_random}
\title{Play a last-owl-farthest strategy}
\usage{
strategy_last_owl_random(game)
}
\arguments{
\item{game}{A current game state}
}
\value{
list with these elements:
  \code{card} string indicating what card to play
  \code{owl} an integer indicating the owl to move
}
\description{
Play a sun if you have a sun, otherwise move the last owl using a random
card.
}
\examples{
game <- setup_game()
strategy_last_owl_random(game)
}
\seealso{
\code{\link{play_card}}
}

