% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_card.R
\name{play_card}
\alias{play_card}
\title{Play a card}
\usage{
play_card(card, owl, game)
}
\arguments{
\item{card}{A character string indicating a card}

\item{owl}{An integer specifying which owl to move. Owl #1 is closest to the
end.}

\item{game}{A game state}
}
\value{
An updated game state
}
\description{
Play a card
}
\examples{
game <- setup()
game <- play(game$player$player1[1])

}
\seealso{
\link{\code{setup}}
}

