% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_experiments.R
\name{run_experiment}
\alias{run_experiment}
\title{Run a Hoot Owl Hoot! experiment}
\usage{
run_experiment(n_players, n_owls, strategy, n_cards_per_player = 3,
  verbosity = 0)
}
\arguments{
\item{n_players}{An integer specifying the number of players}

\item{n_owls}{An integer specifying the number of owls}

\item{strategy}{A function specifying a strategy for all players to use or a
list of length \code{n_players} where element i specifies a strategy for
player i. See details for more information.}

\item{n_cards_per_player}{An integer specifying the number of cards for each
player.}

\item{verbosity}{An integer specifying how much detail to print out (0=none).}
}
\value{
A data.frame with these end game statistics:
  win: logical indicating if the game was won
  n_cards_played: the number of cards played by the players
  n_suns_played: the number of sun cards played by the players (max: 13)
  owl_score: the sum of the space ids for the owls (lower is better, min: 0)
  n_owls_left: the number of owls left on the board
}
\description{
Simulate a Hoot Owl Hoot! game and record whether the game was
a win or loss, how many cards were played, and how many suns were played.
}
\examples{
run_experiment(n_players = 2, n_owls = 2, strategy = strategy_random)

}
\seealso{
\code{\link{simulate_game}}, \code{\link{run_experiments}}
}

